Moved Repo out of own folder and deleted that folder (was unnecessary)
